remove code from legacy course_catalog api
