--- frontends/mit-open/src/page-components/ExpandedLearningResourceDisplay/learning-resource-drawer.scss
@@ -1,265 +1,0 @@
-@use "ol-utilities/assets/breakpoint";
-
-.expanded-lr-summary {
-  flex-direction: column;
-  font-size: 16px;
-  line-height: 1.3;
-  padding: 30px;
-
-  select {
-    background: #f8f8f8;
-    border: 1px solid #bbb;
-  }
-
-  .run-selector {
-    display: inline-flex;
-    font-size: 13px;
-
-    .info-label {
-      color: #b0b0b0;
-      padding-right: 6px;
-    }
-  }
-
-  .ol-lrc-cert {
-    float: right;
-  }
-
-  .image-div {
-    width: 100%;
-    overflow: hidden;
-    margin-top: 10px;
-    margin-bottom: 10px;
-
-    img {
-      display: block;
-      margin: auto;
-      width: 100%;
-    }
-  }
-
-  .link-share-offered-by {
-    margin-top: 20px;
-    margin-bottom: 20px;
-  }
-
-  .object-type {
-    font-size: 15px;
-    font-weight: 600;
-    letter-spacing: 0;
-    line-height: 19px;
-    text-transform: uppercase;
-    color: #b0b0b0;
-  }
-
-  .external-links {
-    display: inline-flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 37px;
-    border-radius: 4px;
-    background-color: #a32034;
-    padding-right: 10px;
-    padding-left: 10px;
-    color: white;
-    text-transform: capitalize;
-    margin: auto;
-    width: 128px;
-    cursor: pointer;
-    float: left;
-    margin-right: 20px;
-
-    a.link-button {
-      margin: auto;
-      font-size: 18px;
-      line-height: 37px;
-    }
-  }
-
-  .section-label {
-    color: black;
-    margin-bottom: 10px;
-    font-weight: bold;
-  }
-
-  .offered-by {
-    display: inline-flex;
-    font-size: 15px;
-    height: 37px;
-    align-items: center;
-
-    .label {
-      color: #b0b0b0;
-    }
-
-    .offeror {
-      font-weight: bold;
-    }
-
-    @include breakpoint.up(sm) {
-      float: right;
-    }
-  }
-
-  .title {
-    font-size: 20px;
-    color: black;
-    font-weight: bold;
-    margin: 5px 0;
-  }
-
-  .description {
-    font-size: 16px;
-    color: black;
-
-    .tt-expansion-control {
-      @include breakpoint.down(sm) {
-        text-align: right;
-        padding-right: 25px;
-        padding-top: 5px;
-        display: block;
-        color: #202020;
-        font-weight: bold;
-      }
-    }
-
-    @include breakpoint.down(sm) {
-      font-size: 14px;
-    }
-  }
-}
-
-.ol-lr-info {
-  dl {
-    display: grid;
-    grid-template-columns: 1fr 2fr;
-    row-gap: 15px;
-  }
-}
-
-.elr-share {
-  display: inline-block;
-  border: 1px solid #d8d8d8;
-  border-radius: 4px;
-  background-color: #fafafa;
-  height: 37px;
-  align-items: center;
-  margin: auto;
-  width: 128px;
-  cursor: pointer;
-  font-size: 18px;
-  margin-right: 20px;
-
-  .share-contents {
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-    font-size: 18px;
-    line-height: 37px;
-
-    i {
-      transform: scaleX(-1);
-      font-size: 30px;
-      margin-left: 23px;
-    }
-  }
-}
-
-$red-link-color: #9a2a33;
-$link-padding: 11px 20px 11px 11px;
-$footer-height: 172px;
-$drawer-trans: width 500ms;
-
-.lr-drawer {
-  .MuiDrawer-paper {
-    width: 100%;
-    max-width: 600px;
-  }
-}
-
-.lr-drawer.MuiDrawer-root .drawer-close {
-  float: right;
-  padding: 20px;
-  padding-bottom: 0;
-
-  .material-icons {
-    font-size: 35px;
-    float: right;
-  }
-
-  border: none;
-  background-color: white;
-}
-
-.drawer-mobile-header {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  height: 60px;
-}
-
-.tooltip {
-  background: white;
-  padding: 12px;
-  min-width: 250px;
-  max-width: 250px;
-  opacity: 1 !important;
-  z-index: 5000;
-
-  .tooltip-text,
-  h4 {
-    width: 90%;
-  }
-
-  h4 {
-    padding: 0 0 10px;
-    margin: 0;
-    font-size: inherit;
-  }
-
-  input {
-    width: 100%;
-    box-sizing: border-box;
-  }
-
-  .bottom-row {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    margin-top: 15px;
-
-    .tooltip-buttons {
-      display: flex;
-
-      > div {
-        margin-right: 3px;
-      }
-    }
-
-    .copy-url {
-      font-size: 12px;
-    }
-  }
-}
-
-.rc-tooltip {
-  opacity: 1;
-  box-shadow: 0 0 20px 0;
-
-  .rc-tooltip-inner {
-    border: none;
-    padding: 0;
-  }
-
-  .rc-tooltip-arrow {
-    border-top-color: white;
-  }
-
-  position: absolute;
-  z-index: 5000;
-  display: block;
-  line-height: 1.5;
-  font-size: 12px;
-  background-color: rgb(0 0 0 / 5%);
-  padding: 1px;
-}
