+++ frontends/mit-open/src/page-components/LearningResourceCard/LearningResourceCard.test.tsx
@@ -1,13 +1,9 @@
 import React from "react"
 import * as NiceModal from "@ebay/nice-modal-react"
-import { makeLearningResource } from "ol-utilities/factories"
 import { renderWithProviders, user, screen } from "../../test-utils"
 import type { User } from "../../test-utils"
 import LearningResourceCard from "./LearningResourceCard"
-import type {
-  LearningResourceCardPropsOld,
-  LearningResourceCardPropsNew,
-} from "./LearningResourceCard"
+import type { LearningResourceCardProps } from "./LearningResourceCard"
 import AddToListDialog from "./AddToListDialog"
 import * as factories from "api/test-utils/factories"
 
@@ -20,46 +16,11 @@ jest.mock("@ebay/nice-modal-react", () => {
   }
 })
 
-describe("LearningResourceCard (old interface)", () => {
-  type SetupOptions = {
-    userSettings?: Partial<User>
-    props?: Partial<LearningResourceCardPropsOld>
-  }
-  const setup = ({ userSettings: user, props = {} }: SetupOptions = {}) => {
-    const { resource = makeLearningResource(), variant = "column" } = props
-    const { view, location } = renderWithProviders(
-      <LearningResourceCard {...props} resource={resource} variant={variant} />,
-      { user },
-    )
-    return { resource, view, location }
-  }
-
-  test("Clicking resource title routes to LearningResourceDrawer", async () => {
-    const { resource, location } = setup()
-    expect(location.current.search).toBe("") // Drawer is closed
-    await user.click(screen.getByRole("heading", { name: resource.title }))
-
-    const actual = new URLSearchParams(location.current.search).sort()
-    const expected = new URLSearchParams(
-      Object.entries({
-        resource_type: resource.object_type,
-        resource_id: String(resource.id),
-      }),
-    ).sort()
-    expect(actual).toEqual(expected)
-  })
-
-  test("Applies className to the resource card", () => {
-    const { view } = setup({ props: { className: "test-class" } })
-    expect(view.container.firstChild).toHaveClass("test-class")
-  })
-})
-
-describe("LearningResourceCard (new interface)", () => {
+describe("LearningResourceCard", () => {
   const makeResource = factories.learningResources.resource
   type SetupOptions = {
     userSettings?: Partial<User>
-    props?: Partial<LearningResourceCardPropsNew>
+    props?: Partial<LearningResourceCardProps>
   }
   const setup = ({ userSettings: user, props = {} }: SetupOptions = {}) => {
     const { resource = makeResource(), variant = "column" } = props
