--- frontends/mit-open/src/page-components/LearningResourceCard/LearningResourceDrawer.tsx
@@ -1,98 +1,0 @@
-import React, { useCallback } from "react"
-import { useResource } from "@/services/api/learning-resources"
-import { RoutedDrawer } from "ol-components"
-import { LearningResource } from "ol-common"
-import ExpandedLearningResourceDisplay, {
-  ExpandedLearningResourceDisplayProps as LRDisplayProps,
-} from "@/page-components/ExpandedLearningResourceDisplay/ExpandedLearningResourceDisplay"
-import { useSearchParams } from "ol-utilities"
-
-const RESOURCE_ID_PARAM = "resource_id"
-const RESOURCE_TYPE_PARAM = "resource_type"
-const RESOURCE_PARAMS = [RESOURCE_ID_PARAM, RESOURCE_TYPE_PARAM] as const
-
-const formatShareLink: LRDisplayProps["formatShareLink"] = (resource) => {
-  const search = new URLSearchParams()
-  search.set(RESOURCE_ID_PARAM, String(resource.id))
-  search.set(RESOURCE_TYPE_PARAM, resource.object_type)
-  return `${window.location.origin}${window.location.pathname}?${search}`
-}
-
-const drawerImg = {
-  ocwBaseUrl: window.SETTINGS.ocw_next_base_url,
-  embedlyKey: window.SETTINGS.embedlyKey,
-  width: 440,
-  height: 239,
-}
-
-export type ResourceIdentifiers = {
-  type: string
-  id: number
-  runId?: number | string | undefined
-}
-
-type DrawerContentProps = {
-  resourceType: string
-  resourceId: number
-}
-const DrawerContent: React.FC<DrawerContentProps> = ({
-  resourceId,
-  resourceType,
-}) => {
-  const resourceQuery = useResource(resourceType, resourceId)
-  const resource = resourceQuery.data
-  return (
-    <section className="ic-lr-drawer" aria-label="Detailed description">
-      {resource && (
-        <ExpandedLearningResourceDisplay
-          imgConfig={drawerImg}
-          resource={resource}
-          formatShareLink={formatShareLink}
-        />
-      )}
-    </section>
-  )
-}
-
-const LearningResourceDrawer: React.FC = () => {
-  return (
-    <RoutedDrawer
-      anchor="right"
-      params={RESOURCE_PARAMS}
-      requiredParams={RESOURCE_PARAMS}
-    >
-      {(childProps) => (
-        <DrawerContent
-          resourceId={Number(childProps.params.resource_id)}
-          resourceType={childProps.params.resource_type}
-        />
-      )}
-    </RoutedDrawer>
-  )
-}
-
-type ActivateResourceDrawer = (
-  e: Pick<LearningResource, "id" | "object_type">,
-) => void
-/**
- * Returns a callback that that updates the URL to activate the
- * LearningResourceDrawer.
- *
- * The callback is stable between re-renders.
- */
-const useActivateResourceDrawer = (): ActivateResourceDrawer => {
-  const [searchParams, setSearchParams] = useSearchParams()
-  return useCallback(
-    (resource) => {
-      const params = new URLSearchParams(searchParams)
-      params.set(RESOURCE_ID_PARAM, String(resource.id))
-      params.set(RESOURCE_TYPE_PARAM, resource.object_type)
-      setSearchParams(params)
-    },
-    [setSearchParams, searchParams],
-  )
-}
-
-export default LearningResourceDrawer
-
-export { useActivateResourceDrawer }
