--- frontends/mit-open/src/page-components/LearningResourceCardTemplateLegacy/learning-resource-card-legacy.scss
@@ -1,150 +1,0 @@
-@use "sass:math";
-
-$light-text: #8c8c8c !default;
-$spacer: 0.75rem !default;
-$bold-weight: bold !default;
-$small-font-size: 0.75em !default;
-
-.ol-lrc-root-old {
-  // Ensure the resource image borders match card borders
-  .MuiCardMedia-root,
-  > .MuiCardContent-root {
-    border-radius: inherit;
-  }
-
-  .ol-lrc-col {
-    display: flex;
-    flex-direction: column;
-  }
-
-  .ol-lrc-row {
-    display: flex;
-    flex-direction: row;
-  }
-
-  .ol-lrc-row-reverse {
-    display: flex;
-    flex-direction: row-reverse;
-  }
-
-  .ol-lrc-row-reverse .ol-lrc-image {
-    margin-left: 16px;
-  }
-
-  .ol-lrc-row .ol-lrc-image {
-    margin-right: 16px;
-  }
-
-  .ol-lrc-content .clickable-title {
-    border: none;
-    background-color: white;
-    display: block;
-    text-align: left;
-    padding: 0;
-    margin: 0;
-  }
-
-  .ol-lrc-content .clickable-title:hover {
-    text-decoration: underline;
-    cursor: pointer;
-  }
-
-  .ol-lrc-details {
-    /*
-    Make content flexbox so that we can control which child fills remaining space.
-    */
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-
-    > * {
-      /*
-      Flexbox doesn't have collapsing margins, so we need to avoid double spacing.
-      The column-gap property would be a nicer solution, but it doesn't have the
-      best browser support yet.
-      */
-      margin-top: math.div($spacer, 2);
-      margin-bottom: math.div($spacer, 2);
-
-      &:first-child {
-        margin-top: 0;
-      }
-
-      &:last-child {
-        margin-bottom: 0;
-      }
-    }
-
-    .ol-lrc-fill-space-content-end {
-      /*
-      Last child of ol-lrc-content will take up any extra space (flex: 1) but
-      with its contents at the bottom of its box.
-      */
-      flex: 1;
-      display: flex;
-      flex-direction: column;
-      justify-content: flex-end;
-      align-items: flex-start; // The default is stretch, we we do not want.
-    }
-  }
-
-  .ol-lrc-type-row {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    min-height: 1.5em; // ensure consistent height even if no certificate
-  }
-
-  .ol-lrc-type {
-    text-transform: uppercase;
-    color: $light-text;
-  }
-
-  .ol-lrc-cert {
-    height: 1.5em;
-  }
-
-  .ol-lrc-offered-by {
-    color: $light-text;
-    padding-right: 0.25em;
-  }
-
-  .ol-lrc-title {
-    font-weight: $bold-weight;
-    margin: 0;
-  }
-
-  .ol-lrc-drag-handle {
-    display: flex;
-    align-items: center;
-    font-size: 40px;
-    align-self: stretch;
-    color: $light-text;
-    border-right: 1px solid $light-text;
-    margin-right: 16px;
-  }
-
-  .MuiCardContent-root.ol-lrc-sortable {
-    padding-left: 4px;
-  }
-
-  .ol-lrc-footer-row {
-    min-height: 2.5 * $small-font-size; // ensure consistent spacing even if no date
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-  }
-
-  .ol-lrc-chip.MuiChip-root {
-    height: 2.5 * $small-font-size;
-    font-size: $small-font-size;
-
-    .MuiSvgIcon-root {
-      height: 1.25 * $small-font-size;
-      width: 1.25 * $small-font-size;
-    }
-  }
-}
