--- frontends/mit-open/src/services/api/fields/index.ts
@@ -1,3 +1,0 @@
-export * from "./hooks"
-export * from "./interfaces"
-export * as urls from "./urls"
