--- frontends/mit-open/src/services/api/fields/urls.ts
@@ -1,5 +1,0 @@
-export const fieldsList = "/fields"
-
-export const DEFAULT_PAGE_SIZE = 50
-
-export const fieldDetails = (name: string) => `/fields/${name}/`
