--- frontends/mit-open/src/services/api/learning-resources/index.ts
@@ -1,4 +1,0 @@
-export * from "./resources"
-export * from "./favorites"
-export * from "./resourceLists"
-export { urls, DEFAULT_PAGINATION_PARAMS } from "./urls"
