--- frontends/mit-open/src/services/api/widgets/index.ts
@@ -1,2 +1,0 @@
-export * from "./hooks"
-export * as urls from "./urls"
