--- frontends/mit-open/src/services/api/widgets/urls.ts
@@ -1,5 +1,0 @@
-import type { WidgetInstance } from "ol-widgets"
-
-type WidgetId = WidgetInstance["id"]
-
-export const widgetList = (id: WidgetId) => `/widget_lists/${id}/`
