--- frontends/ol-utilities/src/deprecated/index.ts
@@ -1,1 +1,0 @@
-export * from "./ol-search-ui-util"
