+++ frontends/ol-utilities/src/factories.ts
@@ -1,3 +1,1 @@
-export * from "./test-utils/factories/ol-search-ui-factories"
-export * from "./test-utils/factories/ol-learning-resources-factories"
 export * from "./test-utils/factories/factories"
