+++ frontends/ol-utilities/src/test-utils/factories/factories.ts
@@ -2,6 +2,7 @@ import { faker } from "@faker-js/faker/locale/en"
 import { times } from "lodash"
 import type { PartialDeep } from "type-fest"
 import type { PaginatedResult } from "./interfaces"
+import type { EmbedlyConfig } from "ol-common"
 
 export type Factory<T, U = never> = (overrides?: Partial<T>, options?: U) => T
 export type PartialFactory<T, U = T> = (overrides?: PartialDeep<T>) => U
@@ -31,3 +32,15 @@ export const makePaginatedFactory =
  * Make a random URL with `faker`, but standardize it to what browsers use.
  */
 export const makeUrl = (): string => new URL(faker.internet.url()).toString()
+
+export const makeImgConfig: Factory<EmbedlyConfig> = (overrides) => {
+  const imgConfig = {
+    width: faker.datatype.number(),
+    height: faker.datatype.number(),
+    key: faker.datatype.uuid(),
+  }
+  return {
+    ...imgConfig,
+    ...overrides,
+  }
+}
